refactor: time.ts 네이밍 변경에 따른 수정사항
