+++ frontend/src/@components/coupon/CouponStatus/index.tsx
@@ -1,6 +1,6 @@
 import theme from '@/styles/theme';
 import { COUPON_STATUS } from '@/types/client/coupon';
-import { generateDateText } from '@/utils';
+import { generateDateText } from '@/utils/time';
 
 import * as Styled from './style';
